mycolony/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
celery-types==0.23.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.5
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-timezone-field==7.1
djangorestframework==3.16.0
firebase_admin==7.0.0
gcloud==0.17.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jws==0.1.3
kombu==5.5.4
msgpack==1.1.1
mysqlclient==2.2.7
oauth2client==3.0.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.4.3
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
Pyrebase==3.0.27
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jwt==2.0.1
redis==5.2.1
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==0.7.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
